flask_backend/requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-JWT-Extended>=4.4.4
Flask-Marshmallow>=0.14.0
Flask-CORS>=4.0.0
marshmallow>=3.19.0
marshmallow-sqlalchemy>=0.29.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
werkzeug>=2.2.0
jsonschema>=4.19.0
python-dateutil>=2.8.2
pytest>=8.2.0
